+++ include/kanji_tools/stats/Stats.h
@@ -13,7 +13,7 @@ public:
   // Command line options must specify one or more files and 'data' class is
   // used to lookup kanji found in files - see HelpMessage in Stats.cpp for more
   // details on command line options.
-  Stats(size_t argc, const char** argv, DataPtr data);
+  Stats(u_int8_t argc, const char** argv, DataPtr data);
 
   Stats(const Stats&) = delete;
   // operator= is not generated since there are const members
